Automatically redirect to OAuth2 page from login (with single provider)
Normally Moodle will show all the OAuth2 provider buttons under the regular
login form. However, we don't really want that because it will confuse people
and is an unnecessary extra step.

Adding ?redirect=0 to the end of the login URL will prevent this behaviour.
That may be useful for administrator accounts and such.